+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - chore(release): bump version to 0.11.6 and update .gitignore
 
+## [2026-07-16 05:49]
+
+- feat(core): add gray and lightgray color methods
+
 ## [2026-06-15 01:11]
 
 - feat(logging): add dual console logging to basicConfig